fix flaot of SoftplusDoubleGradKernel to double

+++ paddle/phi/kernels/activation_grad_kernel.h
@@ -266,8 +266,8 @@ void SoftplusDoubleGradKernel(const Context& dev_ctx,
                               const DenseTensor& x,
                               const DenseTensor& dout,
                               const DenseTensor& ddx,
-                              float beta,
-                              float threshold,
+                              double beta,
+                              double threshold,
                               DenseTensor* dx,
                               DenseTensor* ddout);
 template <typename T, typename Context>
